+++ src/callback/multi.rs
@@ -2,7 +2,11 @@
 //! that are called multiple times, i.e. callbacks of events that occur more
 //! than once per callback.
 
-use crate::{callback, panic::CatchUnwind};
+use crate::{
+    callback,
+    panic::{FutureCatchUnwind, Payload, StreamCatchUnwind},
+};
+use futures::stream::Stream;
 use std::{future::Future, panic, pin::Pin, task};
 
 macro_rules! sync_multi {
@@ -31,7 +35,7 @@ macro_rules! async_multi {
             let future = $callback();
             let notifier = notifier.clone();
             let handler_future = Box::pin(async move {
-                let result = CatchUnwind::new(future).await;
+                let result = FutureCatchUnwind::new(future).await;
                 match result {
                     Ok(data) => notifier.success(data),
                     Err(payload) => notifier.panicked(payload),
@@ -45,8 +49,75 @@ macro_rules! async_multi {
     }};
 }
 
+macro_rules! streaming_multi {
+    ($self:expr, $callback:expr) => {{
+        struct Handler<S>
+        where
+            S: Stream,
+        {
+            inner: StreamCatchUnwind<S>,
+            notifier: Option<callback::shared::Notifier<S::Item>>,
+        }
+
+        impl<S> Handler<S>
+        where
+            S: Stream,
+        {
+            fn new(
+                inner: S,
+                notifier: callback::shared::Notifier<S::Item>,
+            ) -> Self {
+                Self {
+                    inner: StreamCatchUnwind::new(inner),
+                    notifier: Some(notifier),
+                }
+            }
+        }
+
+        impl<S> Stream for Handler<S>
+        where
+            S: Stream,
+        {
+            type Item = ();
+
+            fn poll_next(
+                self: Pin<&mut Self>,
+                ctx: &mut task::Context<'_>,
+            ) -> task::Poll<Option<Self::Item>> {
+                let Self { ref mut notifier, ref mut inner } =
+                    unsafe { self.get_unchecked_mut() };
+
+                let some_notifier = match notifier {
+                    Some(notif) => notif,
+                    None => return task::Poll::Ready(None),
+                };
+                match unsafe { Pin::new_unchecked(inner) }.poll_next(ctx) {
+                    task::Poll::Pending => task::Poll::Pending,
+                    task::Poll::Ready(None) => {
+                        *notifier = None;
+                        task::Poll::Ready(None)
+                    },
+                    task::Poll::Ready(Some(Ok(item))) => {
+                        some_notifier.success(item);
+                        task::Poll::Ready(Some(()))
+                    },
+                    task::Poll::Ready(Some(Err(payload))) => {
+                        some_notifier.panicked(payload);
+                        task::Poll::Ready(Some(()))
+                    },
+                }
+            }
+        }
+
+        let (notifier, inner_listener) = callback::shared::channel();
+        let handler = Box::pin(Handler::new($callback, notifier));
+        let ret = ($self.register_fn)(handler as StreamingCbHandler);
+        (ret, Listener::new(inner_listener))
+    }};
+}
+
 /// The type of synchronous, multi-call callback handlers (i.e. the handler that
-/// calls callbacks): a mutable function.
+/// calls callbacks): a boxed mutable function, a wrapper over callbacks.
 pub type SyncCbHandler<'cb> = Box<dyn FnMut() + 'cb>;
 
 /// The type of futures used in asynchronous, multi-call callback handlers (i.e.
@@ -54,9 +125,15 @@ pub type SyncCbHandler<'cb> = Box<dyn FnMut() + 'cb>;
 pub type AsyncCbHandlerFuture<'fut> = Pin<Box<dyn Future<Output = ()> + 'fut>>;
 
 /// The type of asynchronous, multi-call callback handlers (i.e. the handler
-/// that calls callbacks): a mutable function yielding boxed futures.
+/// that calls callbacks): a boxed mutable function yielding boxed futures, a
+/// wrapper over callbacks.
 pub type AsyncCbHandler<'cb, 'fut> =
     Box<dyn FnMut() -> AsyncCbHandlerFuture<'fut> + 'cb>;
+
+/// The type of streaming asyncrhonous, multi-call callbackhandlers (i.e. the
+/// handler that calls stream callbacks): a boxed stream, a wrapper over
+/// callbacks.
+pub type StreamingCbHandler<'cb> = Pin<Box<dyn Stream<Item = ()> + 'cb>>;
 
 /// Register of multi-call callbacks into an event, where the callback is
 /// syncrhonous (though waiting for the callback to complete is still
@@ -545,7 +622,322 @@ impl<F> AsyncRegister<F> {
     }
 }
 
-/// A handle to a multi-call callback registered in an event.
+/// Register of multi-call callbacks into an event, where the callback is an
+/// asyncrhonous streaming (waiting for the callback to complete is also
+/// asynchronous).
+#[derive(Debug, Clone, Copy)]
+pub struct StreamingRegister<F> {
+    register_fn: F,
+}
+
+impl<F> StreamingRegister<F> {
+    /// Creates a new register using an inner register function that can be used
+    /// only once (though the callback might be called multiple times). Such
+    /// function receives callbacks handlers and register them. Callback
+    /// handlers are register-internal streams that are awaited for when the
+    /// event completes, and then, they await the actual callbacks, i.e. a
+    /// wrapper for the actual callback.
+    pub fn new<'cb, T>(register_fn: F) -> Self
+    where
+        F: FnOnce(StreamingCbHandler<'cb>) -> T,
+    {
+        Self { register_fn }
+    }
+
+    /// Creates a new register using an inner register function that can be used
+    /// multiple times, requiring mutability, however. Such function receives
+    /// callbacks handlers and register them. Callback handlers are
+    /// register-internal streams that are awaited for when the event completes,
+    /// and then, they await the actual callbacks, i.e. a wrapper for the actual
+    /// callback.
+    pub fn new_mut<'cb, T>(register_fn: F) -> Self
+    where
+        F: FnMut(StreamingCbHandler<'cb>) -> T,
+    {
+        Self { register_fn }
+    }
+
+    /// Creates a new register using an inner register function that can be used
+    /// multiple times and does not require mutability. Such function receives
+    /// callbacks handlers and register them. Callback handlers are
+    /// register-internal streams that are awaited for when the event completes,
+    /// and then, they await the actual callbacks, i.e. a wrapper for the actual
+    /// callback.
+    pub fn new_ref<'cb, T>(register_fn: F) -> Self
+    where
+        F: Fn(StreamingCbHandler<'cb>) -> T,
+    {
+        Self { register_fn }
+    }
+
+    /// Registers a callback (a stream) and lets it listen for the target event.
+    /// A listener is returned, and calling `[Listener::next]` yields a
+    /// future that waits for an occurence of the event. Whenever the
+    /// callback returns `None` on [`Stream::poll_next`], the whole listener
+    /// stops and awaiting next events will yield
+    /// [`callback::Error::Cancelled`].
+    ///
+    /// This method consumes the register.
+    ///
+    /// # Examples
+    ///
+    /// ## Dummy Example
+    ///
+    /// ```no_run
+    /// use futures::stream::{Stream, StreamExt};
+    /// use std::{
+    ///     future::Future,
+    ///     pin::Pin,
+    ///     task::{Context, Poll},
+    /// };
+    /// use webio::{callback, task};
+    ///
+    /// fn streaming_multi_event<S>(limit: u32, mut callback: Pin<Box<S>>)
+    /// where
+    ///     S: Stream + ?Sized + 'static,
+    /// {
+    ///     if let Some(new_limit) = limit.checked_sub(1) {
+    ///         task::detach(async move {
+    ///             callback.next().await;
+    ///             task::yield_now().await;
+    ///             streaming_multi_event(new_limit, callback);
+    ///         });
+    ///     }
+    /// }
+    ///
+    /// #[derive(Debug)]
+    /// struct IncrementStream {
+    ///     value: u32,
+    ///     curr_done: bool,
+    /// }
+    ///
+    /// impl Default for IncrementStream {
+    ///     fn default() -> Self {
+    ///         Self { value: 0, curr_done: false }
+    ///     }
+    /// }
+    ///
+    /// impl Stream for IncrementStream {
+    ///     type Item = u32;
+    ///
+    ///     fn poll_next(
+    ///         self: Pin<&mut Self>,
+    ///         _ctx: &mut Context<'_>,
+    ///     ) -> Poll<Option<Self::Item>> {
+    ///         let this = self.get_mut();
+    ///         if this.curr_done {
+    ///             let output = this.value;
+    ///             this.curr_done = false;
+    ///             this.value = output.wrapping_add(1);
+    ///             Poll::Ready(Some(output))
+    ///         } else {
+    ///             this.curr_done = true;
+    ///             Poll::Pending
+    ///         }
+    ///     }
+    /// }
+    ///
+    /// # fn main() {
+    /// # task::detach(async {
+    /// let register = callback::multi::StreamingRegister::new(|callback| {
+    ///     streaming_multi_event(3, callback);
+    /// });
+    ///
+    /// let listener = register.listen(IncrementStream::default());
+    ///
+    /// assert_eq!(listener.next().await.unwrap(), 0);
+    /// assert_eq!(listener.next().await.unwrap(), 1);
+    /// assert_eq!(listener.next().await.unwrap(), 2);
+    /// # });
+    /// # }
+    /// ```
+    pub fn listen<'cb, S>(self, callback: S) -> Listener<S::Item>
+    where
+        F: FnOnce(StreamingCbHandler<'cb>),
+        S: Stream + 'cb,
+    {
+        let (_, listener) = self.listen_returning(callback);
+        listener
+    }
+
+    /// Registers a callback (a stream) and lets it listen for the target event.
+    /// A listener is returned, and calling `[Listener::next]` yields a
+    /// future that waits for an occurence of the event. Whenever the
+    /// callback returns `None` on [`Stream::poll_next`], the whole listener
+    /// stops and awaiting next events will yield
+    /// [`callback::Error::Cancelled`].
+    ///
+    /// This method does not consume the register, but requires mutability.
+    pub fn listen_mut<'cb, S>(&mut self, callback: S) -> Listener<S::Item>
+    where
+        F: FnMut(StreamingCbHandler<'cb>),
+        S: Stream + 'cb,
+    {
+        let (_, listener) = self.listen_mut_returning(callback);
+        listener
+    }
+
+    /// Registers a callback (a stream) and lets it listen for the target event.
+    /// A listener is returned, and calling `[Listener::next]` yields a
+    /// future that waits for an occurence of the event. Whenever the
+    /// callback returns `None` on [`Stream::poll_next`], the whole listener
+    /// stops and awaiting next events will yield
+    /// [`callback::Error::Cancelled`].
+    ///
+    /// This method does not consume the register and does not requires
+    /// mutability.
+    pub fn listen_ref<'cb, S>(&self, callback: S) -> Listener<S::Item>
+    where
+        F: Fn(StreamingCbHandler<'cb>),
+        S: Stream + 'cb,
+    {
+        let (_, listener) = self.listen_ref_returning(callback);
+        listener
+    }
+
+    /// Registers a callback (a stream) and lets it listen for the target event.
+    /// A listener is returned, and calling `[Listener::next]` yields a
+    /// future that waits for an occurence of the event. Whenever the
+    /// callback returns `None` on [`Stream::poll_next`], the whole listener
+    /// stops and awaiting next events will yield
+    /// [`callback::Error::Cancelled`]. The register can also return a value,
+    /// and so, this method returns both the register's return value and the
+    /// listener.
+    ///
+    /// This method consumes the register.
+    ///
+    /// # Examples
+    ///
+    /// ## Dummy Example
+    ///
+    /// ```no_run
+    /// use futures::stream::{Stream, StreamExt};
+    /// use std::{
+    ///     future::Future,
+    ///     pin::Pin,
+    ///     task::{Context, Poll},
+    /// };
+    /// use webio::{callback, task};
+    ///
+    /// fn streaming_multi_event<S>(limit: u32, mut callback: Pin<Box<S>>)
+    /// where
+    ///     S: Stream + ?Sized + 'static,
+    /// {
+    ///     if let Some(new_limit) = limit.checked_sub(1) {
+    ///         task::detach(async move {
+    ///             callback.next().await;
+    ///             task::yield_now().await;
+    ///             streaming_multi_event(new_limit, callback);
+    ///         });
+    ///     }
+    /// }
+    ///
+    /// #[derive(Debug)]
+    /// struct IncrementStream {
+    ///     value: u32,
+    ///     curr_done: bool,
+    /// }
+    ///
+    /// impl Default for IncrementStream {
+    ///     fn default() -> Self {
+    ///         Self { value: 0, curr_done: false }
+    ///     }
+    /// }
+    ///
+    /// impl Stream for IncrementStream {
+    ///     type Item = u32;
+    ///
+    ///     fn poll_next(
+    ///         self: Pin<&mut Self>,
+    ///         _ctx: &mut Context<'_>,
+    ///     ) -> Poll<Option<Self::Item>> {
+    ///         let this = self.get_mut();
+    ///         if this.curr_done {
+    ///             let output = this.value;
+    ///             this.curr_done = false;
+    ///             this.value = output.wrapping_add(1);
+    ///             Poll::Ready(Some(output))
+    ///         } else {
+    ///             this.curr_done = true;
+    ///             Poll::Pending
+    ///         }
+    ///     }
+    /// }
+    ///
+    /// # fn main() {
+    /// # task::detach(async {
+    /// let register = callback::multi::StreamingRegister::new(|callback| {
+    ///     streaming_multi_event(3, callback);
+    ///     "some-ret-string"
+    /// });
+    ///
+    /// let (ret, listener) = register.listen_returning(IncrementStream::default());
+    ///
+    /// assert_eq!(ret, "some-ret-string");
+    /// assert_eq!(listener.next().await.unwrap(), 0);
+    /// assert_eq!(listener.next().await.unwrap(), 1);
+    /// assert_eq!(listener.next().await.unwrap(), 2);
+    /// # });
+    /// # }
+    /// ```
+    pub fn listen_returning<'cb, S, T>(
+        self,
+        callback: S,
+    ) -> (T, Listener<S::Item>)
+    where
+        F: FnOnce(StreamingCbHandler<'cb>) -> T,
+        S: Stream + 'cb,
+    {
+        streaming_multi!(self, callback)
+    }
+
+    /// Registers a callback (a stream) and lets it listen for the target event.
+    /// A listener is returned, and calling `[Listener::next]` yields a
+    /// future that waits for an occurence of the event. Whenever the
+    /// callback returns `None` on [`Stream::poll_next`], the whole listener
+    /// stops and awaiting next events will yield
+    /// [`callback::Error::Cancelled`]. The register can also return a value,
+    /// and so, this method returns both the register's return value and the
+    /// listener.
+    ///
+    /// This method does not consume the register but requires mutability.
+    pub fn listen_mut_returning<'cb, S, T>(
+        &mut self,
+        callback: S,
+    ) -> (T, Listener<S::Item>)
+    where
+        F: FnMut(StreamingCbHandler<'cb>) -> T,
+        S: Stream + 'cb,
+    {
+        streaming_multi!(self, callback)
+    }
+
+    /// Registers a callback (a stream) and lets it listen for the target event.
+    /// A listener is returned, and calling `[Listener::next]` yields a
+    /// future that waits for an occurence of the event. Whenever the
+    /// callback returns `None` on [`Stream::poll_next`], the whole listener
+    /// stops and awaiting next events will yield
+    /// [`callback::Error::Cancelled`]. The register can also return a value,
+    /// and so, this method returns both the register's return value and the
+    /// listener.
+    ///
+    /// This method does not consume the register and does not requires
+    /// mutability.
+    pub fn listen_ref_returning<'cb, S, T>(
+        &self,
+        callback: S,
+    ) -> (T, Listener<S::Item>)
+    where
+        F: Fn(StreamingCbHandler<'cb>) -> T,
+        S: Stream + 'cb,
+    {
+        streaming_multi!(self, callback)
+    }
+}
+
+/// A handle to a multi-call callback registered in an event. Typically, the
+/// [`Listener`] is used with the [`Listener::next`] method for awaiting next
+/// occurences of an event, but it can also be used as a stream.
 #[derive(Debug)]
 pub struct Listener<T> {
     inner: callback::shared::Listener<T>,
@@ -559,6 +951,27 @@ impl<T> Listener<T> {
     /// Creates a future that waits for the next occurence of the event.
     pub fn next<'this>(&'this self) -> ListenNext<'this, T> {
         ListenNext::new(self)
+    }
+}
+
+impl<T> Stream for Listener<T> {
+    type Item = Result<T, Payload>;
+
+    fn poll_next(
+        self: Pin<&mut Self>,
+        ctx: &mut task::Context<'_>,
+    ) -> task::Poll<Option<Self::Item>> {
+        match self.inner.receive() {
+            Some(Ok(output)) => task::Poll::Ready(Some(Ok(output))),
+            Some(Err(callback::Error::Panicked(payload))) => {
+                task::Poll::Ready(Some(Err(payload)))
+            },
+            Some(Err(callback::Error::Cancelled)) => task::Poll::Ready(None),
+            None => {
+                self.inner.subscribe(ctx.waker());
+                task::Poll::Pending
+            },
+        }
     }
 }